refactor: split experiment annotation metric hooks
