--- app/src/pages/dataset/metrics/__tests__/ExperimentAnnotationMetricsGrid.test.tsx
@@ -1,67 +1,0 @@
-import { act, type ReactNode } from "react";
-import { createRoot, type Root } from "react-dom/client";
-import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
-
-import { ExperimentAnnotationMetricsGrid } from "../ExperimentAnnotationMetricsGrid";
-
-vi.mock("@phoenix/components", () => ({
-  Flex: ({ children }: { children: ReactNode }) => children,
-  Loading: () => <div>loading annotations</div>,
-  Text: ({ children }: { children: ReactNode }) => children,
-}));
-
-vi.mock("@phoenix/components/experiment", () => ({
-  BaselineExperimentBadge: () => null,
-}));
-
-vi.mock("@phoenix/components/experiment/SequenceNumberToken", () => ({
-  SequenceNumberToken: () => null,
-}));
-
-vi.mock("@phoenix/components/chart", () => ({
-  ChartPanel: ({ children }: { children: ReactNode }) => children,
-  AnnotationMetricsChart: () => null,
-  AnnotationScoreLabelToggle: () => null,
-  compactCategoryXAxisProps: {},
-  compactYAxisProps: {},
-  getDefaultAnnotationMetricsView: () => "labels",
-  normalizeAnnotationMetrics: () => [],
-}));
-
-vi.mock("@phoenix/components/exception", () => ({
-  ErrorBoundary: ({ children }: { children: ReactNode }) => children,
-}));
-
-vi.mock("../useExperimentAnnotationMetricsData", () => ({
-  useExperimentAnnotationMetricNames: () => {
-    throw new Promise(() => undefined);
-  },
-}));
-
-describe("ExperimentAnnotationMetricsGrid", () => {
-  let container: HTMLDivElement;
-  let root: Root;
-
-  beforeEach(() => {
-    container = document.createElement("div");
-    document.body.appendChild(container);
-    root = createRoot(container);
-  });
-
-  afterEach(() => {
-    act(() => root.unmount());
-    container.remove();
-  });
-
-  it("renders non-annotation grid children while annotations are pending", () => {
-    act(() => {
-      root.render(
-        <ExperimentAnnotationMetricsGrid datasetId="dataset-1">
-          <div>token and error charts</div>
-        </ExperimentAnnotationMetricsGrid>
-      );
-    });
-
-    expect(container.textContent).toContain("token and error charts");
-  });
-});
