+++ app/src/pages/dataset/metrics/__tests__/useExperimentAnnotationMetricData.test.tsx
@@ -1,21 +1,12 @@
 import { act, Suspense } from "react";
 import { createRoot, type Root } from "react-dom/client";
 import { RelayEnvironmentProvider } from "react-relay";
-import {
-  Environment,
-  Network,
-  Observable,
-  RecordSource,
-  Store,
-} from "relay-runtime";
+import { Environment, Network, RecordSource, Store } from "relay-runtime";
 import { afterEach, beforeEach, describe, expect, it } from "vitest";
 
-import {
-  useExperimentAnnotationMetricData,
-  useExperimentAnnotationMetricNames,
-} from "../useExperimentAnnotationMetricsData";
+import { useExperimentAnnotationMetricData } from "../useExperimentAnnotationMetricData";
 
-describe("useExperimentAnnotationMetricNames", () => {
+describe("useExperimentAnnotationMetricData", () => {
   let container: HTMLDivElement;
   let root: Root;
 
@@ -28,46 +19,6 @@ describe("useExperimentAnnotationMetricNames", () => {
   afterEach(() => {
     act(() => root.unmount());
     container.remove();
-  });
-
-  it("loads annotation names without requesting experiment metrics", async () => {
-    const requestedOperations: {
-      name: string;
-      text: string | null | undefined;
-    }[] = [];
-    const environment = new Environment({
-      network: Network.create((operation) => {
-        requestedOperations.push({
-          name: operation.name,
-          text: operation.text,
-        });
-        return Observable.create(() => undefined);
-      }),
-      store: new Store(new RecordSource()),
-    });
-
-    await act(async () => {
-      root.render(
-        <RelayEnvironmentProvider environment={environment}>
-          <>
-            <Suspense fallback={<div>loading annotation names</div>}>
-              <AnnotationNamesConsumer datasetId="dataset-1" />
-            </Suspense>
-          </>
-        </RelayEnvironmentProvider>
-      );
-    });
-
-    expect(requestedOperations).toHaveLength(1);
-    expect(requestedOperations[0]?.name).toBe(
-      "useExperimentAnnotationMetricNamesQuery"
-    );
-    expect(requestedOperations[0]?.text).toContain(
-      "experimentAnnotationSummaries"
-    );
-    expect(requestedOperations[0]?.text).not.toContain("meanScore");
-    expect(requestedOperations[0]?.text).not.toContain("labelFractions");
-    expect(requestedOperations[0]?.text).not.toContain("metricsExperiments");
   });
 
   it("observes baseline replacements and clears through the Dataset link", async () => {
@@ -149,11 +100,6 @@ describe("useExperimentAnnotationMetricNames", () => {
     );
   });
 });
-
-function AnnotationNamesConsumer({ datasetId }: { datasetId: string }) {
-  useExperimentAnnotationMetricNames(datasetId);
-  return <div>annotation names</div>;
-}
 
 function AnnotationMetricsConsumer({ datasetId }: { datasetId: string }) {
   const { baselineExperiment, experiments } = useExperimentAnnotationMetricData(
